Changed all "count" types to long

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/QuestionService.java
@@ -24,6 +24,6 @@ public interface QuestionService {
 
     List<QuestionVotes> findVotesByQuestionId(Long questionId , Long userId , int page);
 
-    int findVotesByQuestionIdCount(Long questionId, Long userId);
+    long findVotesByQuestionIdCount(Long questionId, Long userId);
 
 }
