+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/SearchService.java
@@ -6,12 +6,12 @@ package ar.edu.itba.paw.interfaces.services;
 
 public interface SearchService {
 
-	Integer countQuestionQuery(String query , SearchFilter filter , SearchOrder order , Number community , User user );
+	long countQuestionQuery(String query , SearchFilter filter , SearchOrder order , Number community , User user );
 	List<Question> search(String query , SearchFilter filter , SearchOrder order , Number community ,User user , int limit , int offset);
 	List<User> searchUser(String query , int limit , int offset , String email);
 	List<Community> searchCommunity(String query, int limit , int offset);
-	Integer searchUserCount(String query , String email);
-	Integer searchCommunityCount(String query);
+	long searchUserCount(String query , String email);
+	long searchCommunityCount(String query);
 	//Te da las preguntas recientes de gente con buen karma
 
 	List<Community> searchCommunity(String query, AccessType accessType , Integer moderatorId , Integer userId, int limit , int offset);
